+++ firestore-bigquery-export/scripts/gen-schema-view/jest.config.js
@@ -14,4 +14,15 @@ module.exports = {
     "<rootDir>/src/__tests__/bigquery/*.test.ts",
     "<rootDir>/src/__tests__/schema-loader-utils/*.test.ts",
   ],
+  moduleNameMapper: {
+    "firebase-admin/eventarc":
+      "<rootDir>/node_modules/firebase-admin/lib/eventarc",
+    "firebase-functions/v2": "<rootDir>/node_modules/firebase-functions/lib/v2",
+    "firebase-admin/auth": "<rootDir>/node_modules/firebase-admin/lib/auth",
+    "firebase-admin/app": "<rootDir>/node_modules/firebase-admin/lib/app",
+    "firebase-admin/database":
+      "<rootDir>/node_modules/firebase-admin/lib/database",
+    "firebase-admin/firestore":
+      "<rootDir>/node_modules/firebase-admin/lib/firestore",
+  },
 };